+++ selftests/check.py
@@ -27,9 +27,9 @@ TEST_SIZE = {
     "job-api-check-tmp-directory-exists": 1,
     "nrunner-interface": 90,
     "nrunner-requirement": 28,
-    "unit": 699,
+    "unit": 697,
     "jobs": 11,
-    "functional-parallel": 316,
+    "functional-parallel": 318,
     "functional-serial": 7,
     "optional-plugins": 0,
     "optional-plugins-golang": 2,
