+++ selftests/unit/runner_asset.py
@@ -1,5 +1,4 @@
 import unittest
-from unittest.mock import patch
 
 from avocado.core.nrunner.runnable import Runnable
 from avocado.plugins.runners.asset import AssetRunner
@@ -34,51 +33,6 @@ class BasicTests(unittest.TestCase):
                 break
         self.assertEqual(messages[-1]["result"], "error")
         stderr = b"Failed to fetch foo ("
-        self.assertIn(stderr, messages[-2]["log"])
-
-
-class FetchTests(unittest.TestCase):
-    """Unit tests for the actions on RequirementPackageRunner class"""
-
-    def setUp(self):
-        """Mock SoftwareManager"""
-
-        self.asset_patcher = patch("avocado.plugins.runners.asset.Asset", autospec=True)
-        self.mock_asset = self.asset_patcher.start()
-        self.addCleanup(self.asset_patcher.stop)
-
-    def test_success_fetch(self):
-
-        self.mock_asset.return_value.fetch.return_value = "/tmp/asset.txt"
-        runnable = Runnable(kind="asset", uri=None, **{"name": "asset.txt"})
-        runner = AssetRunner()
-        status = runner.run(runnable)
-        messages = []
-        while True:
-            try:
-                messages.append(next(status))
-            except StopIteration:
-                break
-        self.assertEqual(messages[-1]["result"], "pass")
-        stdout = b"File fetched at /tmp/asset.txt"
-        self.assertIn(stdout, messages[-3]["log"])
-
-    def test_fail_fetch(self):
-
-        self.mock_asset.return_value.fetch = lambda: (_ for _ in ()).throw(
-            OSError("Failed to fetch asset.txt")
-        )
-        runnable = Runnable(kind="asset", uri=None, **{"name": "asset.txt"})
-        runner = AssetRunner()
-        status = runner.run(runnable)
-        messages = []
-        while True:
-            try:
-                messages.append(next(status))
-            except StopIteration:
-                break
-        self.assertEqual(messages[-1]["result"], "error")
-        stderr = b"Failed to fetch asset.txt"
         self.assertIn(stderr, messages[-2]["log"])
